requirements.txt
psycopg2-binary>=2.9,<3.0
pydantic>=2.7,<3.0
pydantic-settings>=2.1,<3.0
SQLAlchemy>=2.0,<3.0
alembic>=1.13,<2.0
fastapi>=0.115,<1.0
uvicorn[standard]>=0.29,<1.0
python-multipart>=0.0.9,<0.1
opencv-python-headless>=4.9,<5.0
prometheus-client>=0.20,<1.0
aiohttp>=3.9,<4.0
ffmpeg-python>=0.2,<0.3
flake8>=7.0,<8.0
pytest>=8.0,<9.0
pytest-asyncio>=0.23,<0.24
httpx>=0.27,<0.28
